requirements.txt
absl-py==2.0.0
accelerate==0.25.0
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
beautifulsoup4==4.12.2
bert-extractive-summarizer==0.10.1
bitsandbytes==0.40.2
bleach==6.1.0
blessed==1.20.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
comm==0.2.0
commonmark==0.9.1
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.6
dataclasses-json==0.6.3
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distro==1.8.0
et-xmlfile==1.1.0
executing==2.0.1
faiss-cpu==1.7.4
fastjsonschema==2.19.0
filelock==3.9.0
fonttools==4.47.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
getdaft==0.2.7
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gpustat==1.1.1
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kaggle==1.5.16
kiwisolver==1.4.5
langcodes==3.3.0
loralib==0.1.2
ludwig==0.8.6
lxml==4.9.3
mapply==0.1.23
Markdown==3.5.1
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-dataclass==8.5.4
marshmallow-jsonschema==0.13.0
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.13.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.1
nvidia-ml-py==12.535.133
oauthlib==3.2.2
openai==1.5.0
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.1
peft==0.7.1
pexpect==4.9.0
Pillow==9.3.0
platformdirs==4.1.0
pox==0.3.3
ppft==1.7.6.7
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
Pygments==2.17.2
pyparsing==3.1.1
pyrsistent==0.20.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3.post1
pyxlsb==1.0.10
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.4.4
rpds-py==0.15.2
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
terminado==0.18.0
text-unidecode==1.3
thinc==8.2.2
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.15.0
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchdata==0.7.0
torchinfo==1.8.0
torchmetrics==0.11.4
torchtext==0.16.0
torchvision==0.16.0+cu121
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.36.2
triton==2.1.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
wasabi==1.1.2
wcwidth==0.2.12
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
xlrd==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
yarl==1.9.4
